!fixup KTOR-1129 Delegate to original OutgoingContent when logging body
