+++ ktor-client/ktor-client-features/ktor-client-logging/common/src/io/ktor/client/features/logging/LoggedContent.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2014-2020 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package io.ktor.client.features.logging
+
+import io.ktor.http.*
+import io.ktor.http.content.*
+import io.ktor.util.*
+import io.ktor.utils.io.*
+
+internal class LoggedContent(
+    private val originalContent: OutgoingContent,
+    private val channel: ByteReadChannel
+) : OutgoingContent.ReadChannelContent() {
+
+    override val contentType: ContentType? = originalContent.contentType
+    override val contentLength: Long? = originalContent.contentLength
+    override val status: HttpStatusCode? = originalContent.status
+    override val headers: Headers = originalContent.headers
+
+    override fun <T : Any> getProperty(key: AttributeKey<T>): T? = originalContent.getProperty(key)
+
+    override fun <T : Any> setProperty(key: AttributeKey<T>, value: T?) =
+        originalContent.setProperty(key, value)
+
+    override fun readFrom(): ByteReadChannel = channel
+}
