+++ ktor-client/ktor-client-features/ktor-client-logging/common/src/io/ktor/client/features/logging/ObservingUtils.kt
@@ -4,7 +4,6 @@
 
 package io.ktor.client.features.logging
 
-import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.util.*
 import io.ktor.utils.io.*
@@ -21,36 +20,18 @@ internal suspend fun OutgoingContent.observe(log: ByteWriteChannel): OutgoingCon
         val content = readFrom()
 
         content.copyToBoth(log, responseChannel)
-        LoggingContent(this, responseChannel)
+        LoggedContent(this, responseChannel)
     }
     is OutgoingContent.WriteChannelContent -> {
         val responseChannel = ByteChannel()
         val content = toReadChannel()
         content.copyToBoth(log, responseChannel)
-        LoggingContent(this, responseChannel)
+        LoggedContent(this, responseChannel)
     }
     else -> {
         log.close()
         this
     }
-}
-
-internal class LoggingContent(
-    private val originalContent: OutgoingContent,
-    private val channel: ByteReadChannel
-) : OutgoingContent.ReadChannelContent() {
-
-    override val contentType: ContentType? = originalContent.contentType
-    override val contentLength: Long? = originalContent.contentLength
-    override val status: HttpStatusCode? = originalContent.status
-    override val headers: Headers = originalContent.headers
-
-    override fun <T : Any> getProperty(key: AttributeKey<T>): T? = originalContent.getProperty(key)
-
-    override fun <T : Any> setProperty(key: AttributeKey<T>, value: T?) =
-        originalContent.setProperty(key, value)
-
-    override fun readFrom(): ByteReadChannel = channel
 }
 
 private fun OutgoingContent.WriteChannelContent.toReadChannel(
